Address review comment
Namespace helper module inside Killbill
Use yaml to load and generate config
